requirements.txt
sanic==0.6.0
flask==0.12.1
validators==0.11.3
neomodel==3.2.5
bcrypt==3.1.3
pyjwt==1.5.3
# https://pypi.python.org/pypi/py2neo/4.0.0b2
py2neo==3.1.2